Added some comments and moved MixinBase to its own file
