+++ apps/demo/src/app/app.component.html
@@ -81,6 +81,9 @@
     </mat-form-field>
 
     <h2>Input with custom ErrorStateMatcher</h2>
+    <p>An ErrorStateMatcher defines, when a control displays the error state. A custom ErrorStateMatcher can for example be used to display validations on untouched controls.
+      ErrorStateMatchers can either be defined globally or for single controls (seen in the following example).
+    </p>
 
     <app-code-sample [code]="errorStateTs"></app-code-sample>
     <app-code-sample [code]="errorState"></app-code-sample>
