+++ libs/material-file-input/src/lib/file-input/file-input-mixin.ts
@@ -1,0 +1,19 @@
+import { FormGroupDirective, NgControl, NgForm } from '@angular/forms';
+import { CanUpdateErrorStateCtor, ErrorStateMatcher, mixinErrorState } from '@angular/material';
+
+// Boilerplate for applying mixins to FileInput
+/** @docs-private */
+export class FileInputBase {
+  constructor(public _defaultErrorStateMatcher: ErrorStateMatcher,
+      public _parentForm: NgForm,
+      public _parentFormGroup: FormGroupDirective,
+      public ngControl: NgControl) { }
+}
+
+/**
+ * Allows to use a custom ErrorStateMatcher with the file-input component
+ */
+export const FileInputMixinBase:
+  CanUpdateErrorStateCtor &
+  typeof FileInputBase =
+  mixinErrorState(FileInputBase);
